Prepare the documentation for release
The documentation and the examples have bee n updated in an attempt to
provide enough information to anyone who would like to start using this
crate.

Fixes rgreinho/trauma#7
