+++ .github/workflows/ci-rust.yml
@@ -99,62 +99,10 @@ jobs:
       - format
       - lint
       - test
-    if: startsWith(github.ref, 'refs/tags/')
-    strategy:
-      matrix:
-        os:
-          - macos-latest
-          - ubuntu-latest
-          - windows-latest
-        include:
-          - os: macos-latest
-            target: x86_64-apple-darwin
-            extension: ''
-          - os: macos-latest
-            target: aarch64-apple-darwin
-            extension: ''
-          - os: ubuntu-latest
-            target: x86_64-unknown-linux-gnu
-            extension: ''
-          - os: windows-latest
-            target: x86_64-pc-windows-msvc
-            extension: '.exe'
-    runs-on: ${{ matrix.os }}
-
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-          target: ${{ matrix.target }}
-          override: true
-      - uses: Swatinem/rust-cache@v1
-      - name: Build in release mode
-        uses: actions-rs/cargo@v1
-        with:
-          command: build
-          args: --release --all-features --target ${{ matrix.target }}
-      - name: Rename binary
-        shell: bash
-        run: |
-          cp target/${{ matrix.target }}/release/retrieve${{ matrix.extension }} retrieve-${{ matrix.target }}${{ matrix.extension }}
-      - name: Upload artifact
-        uses: actions/upload-artifact@v2
-        with:
-          name: retrieve-${{ matrix.target }}${{ matrix.extension }}
-          path: retrieve-${{ matrix.target }}${{ matrix.extension }}
-
-  publish:
-    needs:
-      - release
     runs-on: ubuntu-latest
     if: startsWith(github.ref, 'refs/tags/')
     steps:
       - uses: actions/checkout@v3
-      - name: Download all workflow run artifacts
-        uses: actions/download-artifact@v2
-      - name: Display structure of downloaded files
-        run: ls -R
       - name: Get Changelog Entry
         id: changelog_reader
         uses: mindsers/changelog-reader-action@v2
@@ -164,4 +112,3 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
           body: ${{ steps.changelog_reader.outputs.changes }}
-          files: retrieve-*
